updating julia wrappers to use 8 byte integers

+++ julia/src/FMM3D.jl
@@ -50,7 +50,7 @@ function lower_level_routs() end
 # fortran input/return types
 
 Fd = Ref{Float64}
-Fi = Ref{Int32}
+Fi = Ref{Int64}
 Fc = Ref{ComplexF64}
 
 # common input types
